requirements.txt
Django==5.1
djangorestframework==3.15.2
openai==1.40.1
